Create 2024-04-05-week-13.md
It’s important to invest your time wisely, as you can’t make more of it. Think about how you typically spend your week as a Develop Carolina Apprentice. Do you feel in control of your schedule, and are you happy with it? Is there time in your day that you’d like to invest differently?